refactor(runtime): fold cloudflare agents into cloudflare platform
