+++ packages/runtime/src/platform/cloudflare/agents/fiber.ts
@@ -1,5 +1,5 @@
-import { drainAgentTurnWithBudget } from "../cloudflare/alarm/run-drain";
-import type { CloudflareDurableObjectStorage } from "../cloudflare/host/durable-object-host";
+import { drainAgentTurnWithBudget } from "../alarm/run-drain";
+import type { CloudflareDurableObjectStorage } from "../host/durable-object-host";
 import { cloudflareAgentsDrainOptionsForPayload } from "./drain-options";
 import {
   type CloudflareAgentsFiberPayload,
