+++ packages/runtime/src/platform/cloudflare/agents/payload-validation.ts
