+++ packages/runtime/src/platform/cloudflare/agents/retry-scheduler.ts
@@ -1,8 +1,8 @@
 import {
   prepareScheduledNotificationRetry,
   shouldRetryNotClaimableScheduledRun,
-} from "../cloudflare/alarm/scheduled-work-context";
-import type { CloudflareDurableObjectStorage } from "../cloudflare/host/durable-object-host";
+} from "../alarm/scheduled-work-context";
+import type { CloudflareDurableObjectStorage } from "../host/durable-object-host";
 import {
   type CloudflareAgentsFiberPayload,
   type CloudflareAgentsThreadFiberPayload,
