+++ packages/runtime/tsdown.config.ts
@@ -3,7 +3,6 @@ import { defineConfig } from "tsdown";
 export default defineConfig({
   entry: [
     "src/index.ts",
-    "src/platform/cloudflare-agents/index.ts",
     "src/platform/cloudflare/index.ts",
     "src/platform/file/index.ts",
     "src/platform/memory/index.ts",
